+++ src/lib/schemas/index.ts
@@ -1,0 +1,9 @@
+import { config } from './config'
+import { profile_sip } from './profile-sip'
+import { profile_bot_discord } from './profile-bot-discord'
+
+export default {
+	config,
+	profile_sip,
+	profile_bot_discord
+}